Fix: User file bugs with babel
User had the same name of a folder User in the same directory and babel didn't like it :(
